+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
   - Include `limit` and `offset` in `thoth-client`'s works query
+  - Paginate `get_works` requests in export API
 
 ### Added
   - Work count query to `thoth-client`